python/requirements.txt
python-simple-hipchat==0.3.3

